Rename files that provide kickstart specifications
Every kickstart module has only one kickstart specification, so the
files can be always called kickstart.py.
