+++ pyanaconda/modules/bar/kickstart.py
