+++ pyanaconda/modules/network/kickstart.py
